+++ src/components/lista-baldes/index.tsx
@@ -1,4 +1,4 @@
-import { useContext } from 'react'
+import { useContext, useMemo } from 'react'
 import { BaldeContext } from '../../context/BaldeContext'
 import  TrashIcon from '../../assets/trash.svg' 
 import  TrashBlack from '../../assets/trash-black.svg' 
@@ -8,12 +8,20 @@ import { toast } from 'react-toastify'
 export const ListaBaldes = () => {
 
 	const { baldes, removeBalde, removerFrutaDoBalde, somarValoresDasFrutas, calcularPercentualOcupacao } = useContext(BaldeContext)
+	
+	const baldesOrdenados = useMemo(() => {
+		return [...baldes].sort((a, b) => {
+			const ocupacaoA = calcularPercentualOcupacao(a.id) || 0
+			const ocupacaoB = calcularPercentualOcupacao(b.id) || 0
+			return ocupacaoB - ocupacaoA
+		})
+	}, [baldes, calcularPercentualOcupacao])
 
-	const handleDelete = (index: number) => {
-
-		const balde = baldes[index]
-		if (balde.frutas.length === 0) {
-			removeBalde(index)
+	
+	const handleDelete = (id: number) => {
+		const balde = baldes.find(balde => balde.id === id)
+		if (balde && balde.frutas.length === 0) {
+			removeBalde(id)
 		} else {
 			toast.warning('O balde não pode ser excluído porque possui frutas adicionadas.')
 		}
@@ -23,17 +31,18 @@ export const ListaBaldes = () => {
 		removerFrutaDoBalde(idFruta, idBalde)
 	}
 
+
 	return (
 		<div className='container-baldes'>
 			<h1>Lista de Baldes</h1>
 			{baldes.length === 0 ? (
 				<p>Não existe nenhum balde :(</p>
 			) : (
-				baldes.map((balde, index) => (
+				baldesOrdenados.map((balde) => (
 					<div key={balde.id}>
 						<h3>
               Capacidade {balde.capacidade}  
-							<button onClick={() => handleDelete(index)}>
+							<button onClick={() => handleDelete(balde.id)}>
 								<img src={TrashIcon} alt="Trash Icon" />
 							</button>
 						</h3>